[NIF] Inheritance Improvements
For NIFs < 10.1.0.114 there are 4 bytes at the front of each block (Group ID), so to parse them the NiObject inheritance would need to be enforced for all blocks since the conditional behavior is on NiObject.   Created a #define in order to easily tag structs as being directly inherited from NiObject.

Also uncapitalized various struct types as to not confuse them with a define.