[NIF] Major RTTI string reorganization
Split the RTTI handling into multiple files.  This organization not only documents which blocks are official and which are not,  it allows several things to be done such as skipping nearly half of the RTTI strings for non-Bethesda NIFs.

Created a separate function for other custom RTTI, and this will eventually grow in the future.  Handling old NetImmerse RTTI separately is planned as well.

FYI the `//` being indented less in the switch/case is for code folding in VSCode.