[NIF] Legacy (< 10.1) support
Kept the NetImmerse and Gamebryo blocks as decoupled as possible.  Aside from a few version condition additions, override blocks were defined in NetImmerse.bt and the overrides are loaded in RTTI.bt for the legacy versions.

In some major inherited blocks such as NiObjectNET or NiAVObject, the override occurs at the beginning of the structs in Objects.bt and short circuits.