[NIF] FO4 precombined extra data
BSPackedCombinedGeomDataExtra is the format generated by the CK and not currently in nif.xml.  It contains the actual geometry and is much larger than BSPackedCombinedSharedGeomDataExtra, which is the format that ships with the vanilla files.  The vanilla files are "shared" because they simply refer to the geometry instead of include it.